Removed a couple of unused classes and moved one.
