+++ rifDataLoaderTool/src/main/java/org/sahsu/rif/dataloader/datastorage/ms/MSSQLSampleRIFDatabaseCreationManager.java
@@ -14,7 +14,7 @@ import org.sahsu.rif.dataloader.concepts.WorkflowState;
 import org.sahsu.rif.dataloader.system.RIFDataLoaderToolError;
 import org.sahsu.rif.dataloader.system.RIFDataLoaderToolMessages;
 import org.sahsu.rif.generic.datastorage.SQLGeneralQueryFormatter;
-import org.sahsu.rif.generic.datastorage.common.SQLQueryUtility;
+import org.sahsu.rif.generic.datastorage.SQLQueryUtility;
 import org.sahsu.rif.generic.datastorage.pg.PGSQLCreateTableQueryFormatter;
 import org.sahsu.rif.generic.datastorage.pg.PGSQLDeleteTableQueryFormatter;
 import org.sahsu.rif.generic.system.RIFGenericLibraryError;
