+++ rifDataLoaderTool/src/main/java/org/sahsu/rif/dataloader/datastorage/pg/PGSQLDataSetManager.java
@@ -11,7 +11,7 @@ import org.sahsu.rif.dataloader.system.RIFDataLoaderToolError;
 import org.sahsu.rif.dataloader.system.RIFDataLoaderToolMessages;
 import org.sahsu.rif.generic.datastorage.DeleteRowsQueryFormatter;
 import org.sahsu.rif.generic.datastorage.SQLGeneralQueryFormatter;
-import org.sahsu.rif.generic.datastorage.common.SQLQueryUtility;
+import org.sahsu.rif.generic.datastorage.SQLQueryUtility;
 import org.sahsu.rif.generic.datastorage.pg.PGSQLInsertQueryFormatter;
 import org.sahsu.rif.generic.datastorage.pg.PGSQLRecordExistsQueryFormatter;
 import org.sahsu.rif.generic.datastorage.pg.PGSQLSelectQueryFormatter;
