+++ rifDataLoaderTool/src/main/java/org/sahsu/rif/dataloader/datastorage/pg/PGSQLDatabaseSetupUtility.java
@@ -16,7 +16,7 @@ import org.sahsu.rif.dataloader.concepts.DatabaseConnectionsConfiguration;
 import org.sahsu.rif.dataloader.system.RIFDataLoaderToolError;
 import org.sahsu.rif.dataloader.system.RIFDataLoaderToolMessages;
 import org.sahsu.rif.generic.datastorage.SQLGeneralQueryFormatter;
-import org.sahsu.rif.generic.datastorage.common.SQLQueryUtility;
+import org.sahsu.rif.generic.datastorage.SQLQueryUtility;
 import org.sahsu.rif.generic.datastorage.pg.PGSQLCreateDatabaseQueryFormatter;
 import org.sahsu.rif.generic.datastorage.pg.PGSQLDropDatabaseQueryFormatter;
 import org.sahsu.rif.generic.system.ClassFileLocator;
