--- rifGenericLibrary/src/main/java/org/sahsu/rif/generic/datastorage/common/package-info.java
@@ -1,9 +1,0 @@
-/**
- * 
- */
-/**
- * Kevin Garwood
- * @author kgarwood
- *
- */
-package org.sahsu.rif.generic.datastorage.common;
