+++ rifGenericLibrary/src/main/java/org/sahsu/rif/generic/datastorage/ms/MSUserDatabaseConnections.java
@@ -11,7 +11,7 @@ import java.util.Set;
 
 import org.sahsu.rif.generic.system.RIFServiceException;
 import org.sahsu.rif.generic.concepts.User;
-import org.sahsu.rif.generic.datastorage.common.SQLQueryUtility;
+import org.sahsu.rif.generic.datastorage.SQLQueryUtility;
 import org.sahsu.rif.generic.system.Messages;
 import org.sahsu.rif.generic.system.RIFGenericLibraryError;
 import org.sahsu.rif.generic.util.RIFLogger;
