+++ rifServices/src/main/java/org/sahsu/rif/services/datastorage/common/RIFContextManager.java
@@ -11,7 +11,7 @@ import org.sahsu.rif.generic.datastorage.AggregateValueQueryFormatter;
 import org.sahsu.rif.generic.datastorage.DatabaseType;
 import org.sahsu.rif.generic.datastorage.RecordExistsQueryFormatter;
 import org.sahsu.rif.generic.datastorage.SelectQueryFormatter;
-import org.sahsu.rif.generic.datastorage.common.SQLQueryUtility;
+import org.sahsu.rif.generic.datastorage.SQLQueryUtility;
 import org.sahsu.rif.generic.system.RIFServiceException;
 import org.sahsu.rif.generic.util.RIFLogger;
 import org.sahsu.rif.services.concepts.AbstractRIFConcept.ValidationPolicy;
