+++ rifServices/src/main/java/org/sahsu/rif/services/datastorage/pg/PGSQLGenerateResultsSubmissionStep.java
@@ -6,7 +6,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 
 import org.sahsu.rif.generic.datastorage.FunctionCallerQueryFormatter;
-import org.sahsu.rif.generic.datastorage.common.SQLQueryUtility;
+import org.sahsu.rif.generic.datastorage.SQLQueryUtility;
 import org.sahsu.rif.generic.system.RIFServiceException;
 import org.sahsu.rif.generic.util.RIFLogger;
 import org.sahsu.rif.services.datastorage.common.GenerateResultsSubmissionStep;
